MINOR: .pyi files taken away.
I finally decided to not use the .pyi files at all as it seems there are some problems with the IDE type annotation when using .py with .pyi files. You necessarly have to define all the different classes, methods and functions inside the .pyi if you want the static type checker to work properly.