MIDDLE: class for annotation of the arc-length on a curce. Corresponding test code also exists.
I created a new class to be able to easily annotate the arc-length of a curve on the curve itself. A corresponding test code was also added. The results seems fine but there are still some problems in the annotation text angle. Will need to test it on real data to try and find a solution that always works.