MIDDLE: 10 concurrent SSH connections max.
I updated the code so that now, when fetching data from the server using the SSHMirroredFilesystem class, there is a maximum of 10 connections that try to fetch a file at the same time. This was added because the SSH control socket file has an upper limit of 10 fetches at the same time. Tested the code and it works perfectly.