MINOR: update (or downgrade) of the .running_time decorator.
I have reverted back the running_time decorator so that when used as a decorator factory there is only the flush: bool argument. In practice, there is no point for it to check all class attributes to see if flush and verbose exist. Waste of time. Also, I moved the method that converted the time in seconds to human readible format to the formatter code.