MAJOR: ProcessAllocator and CustomManager input stack update.
I updated the stack so that when you submit a group of tasks, the arguments should be a generator callable. This is done so that the input stack only has generators inside it and duplicates/creates the arguments only when called by a worker. Saves up a lot of RAM and computing power as a fraction of the data is serialised to initially get to the manager.