MIDDLE: ProcessAllocator changes in Locks.
Not a lot of code changes but should be major for the processing speed depending on the configuration. Before, all pickling operations (even those inside the manager and processes) were on the same lock. I think it was the main reason for the processors blocking when a lot of tasks finished at the same time. Now, the pickling on the server side and for the processsors can be done at the same time. Need to check the difference in processing speed for the spice-line-fits code that I am testing the ProcessAllocator on.