MINOR: ProcessAllocator no results return implimentation.
I added the possibility for the user to decide that the specific tasks sent will not return any results. As such the code will not wait for the results to be joined together nor send them to the manager. The corresponding overloads were also added so that the user knows that the return of the submit is None (as opposed to the task id needed to find the result).