MAJOR: ManagerAllocator class added.
I created an incorporate a ManagerAllocator class inside the ProcessAllocator class. This was done so that the user can decide how many managers he wants for sending the tasks to the workers and for managers that will sort and give the results from the multiprocessing. Seems to work well. Need to test on the spice-line-fits code to see if it works there and if there is any noticable processing speed improvements. The docstrings weren't added nor updated yet.