MINOR: improvements in sorter allocations for the ManagerAllocator code.
In my last version, the returns from the multiprocessing that had indexes close to each other were put in the same sorter. This could lead to problems as tasks with close indexes might be sent to the sorter around the same time, hence blocking the sorters. Now it just goes from the first sorter to the last in order.