MIDDLE: in ManagerAllocator, simple int counters changed to mp.Values.
I changed the int counts to mp.Values as they are shared memory objects and hence it bypasses the slow manager connectons and even the pickling (the piclking time for an int is really low but the manager connection is really slow so it is worth the change). To do so, I decided to use a singleton to pass those values. That being said, most likely this won't work on Windows OS due to the re-imports done when spawning a process.