MINOR: ProcessAllocator shared memory closing error fix.
I fixed the error at the end of the code where closing the shared memory gave an error saying that the pointer is still being used. Furthermore, I quickly checked the docstring of the codes related to ProcessAllocator and they seem to be all cleaned up. Code now runs properly. Still need to check the toml files setup.