MAJOR: ProcessAllocator changed to a singleton.
While not much code has changed, the improvements are really important as the usage and the efficiency of the code (especially for nested multiprocessing) has improved. ProcessAllocator now being a singleton means that to submit tasks in a nested multiprocessing context, there is no more need to pass around the instance to the class. As a singleton, just calling the class will return the main instance to said class. Also, the serialised class instance is not passed around through the queues anymore. Lastly, it also means that when using the process allocator, one can always call process allocator to 'create' a new instance as, as long as an instance was already created, then it will just reuse that instance.