MIDDLE: ProcessAllocator update: number of tasks can be unrelated to the list length for the 'different_kwargs' parameter.
Now, you can decide to set the number of tasks (for a submission to the process allocator) to a number different than the length of the lists inside the 'different_kwargs' argument. This is so less tasks pass through the managers and for the case where the target function for the tasks needs to take a slice (and not just a unique item) from each list specified in the 'different_kwargs' argument. Seems to work properly.