MINOR: Important error found and fixed (?).
I found an important error in index calculations in the ProcessAllocator related codes. Still seems to be other problems but not sure if it comes from the ProcessAllocator related codes or from the changes I made to Plowman's code. Need further testing (committing to be able to run it at home too).