NULL
Took away the prints for testing as the error was found (I was using a lambda function inside plowman's code. Lambda functions are not picklable). Kept the try except for now as I think that there is an error somewhere inside the ProcessAllocator code. Might have already fixed it though.